local_requirements.txt
# Django/Framework Packages
# Ironwood

appdirs==1.4.3            # via fs
backports.os==0.1.1       # via fs
certifi==2019.9.11        # via requests
chardet==3.0.4            # via requests
django-crum==0.7.3
django-ipware==2.1.0
django-model-utils==3.2.0
django-waffle==0.17.0
django-webpack-loader==0.6.0
django==1.11.25
djangorestframework-jwt==1.11.0  # via edx-drf-extensions
djangorestframework==3.9.4
edx-django-utils==2.0.0   # via edx-drf-extensions
edx-drf-extensions==2.4.0
edx-opaque-keys==2.0.0
edx-rest-api-client==1.9.2
edx-when==0.5
enum34==1.1.6             # via fs
event-tracking==0.2.9
fs==2.4.11                # via xblock
future==0.17.1            # via backports.os, pyjwkest
idna==2.8                 # via requests
jsonfield==2.0.2
lxml==4.4.1               # via xblock
markupsafe==1.1.1         # via xblock
newrelic==5.2.0.127       # via edx-django-utils
pbr==5.4.3                # via stevedore
psutil==1.2.1             # via edx-django-utils, edx-drf-extensions
pycryptodomex==3.9.0
pyjwkest==1.3.2           # via edx-drf-extensions
pyjwt==1.7.1              # via djangorestframework-jwt, edx-rest-api-client
pymongo==3.9.0            # via edx-opaque-keys, event-tracking
python-dateutil==2.8.0
pytz==2019.2
pyyaml==5.1.2             # via xblock
requests==2.22.0          # via edx-drf-extensions, edx-rest-api-client, pyjwkest, slumber
rest-condition==1.0.3     # via edx-drf-extensions
rules==2.1
semantic-version==2.8.2   # via edx-drf-extensions
six==1.12.0               # via edx-drf-extensions, edx-opaque-keys, event-tracking, fs, pyjwkest, python-dateutil, stevedore, xblock
slumber==0.7.1            # via edx-rest-api-client
stevedore==1.31.0         # via edx-opaque-keys
typing==3.7.4.1           # via fs
urllib3==1.25.6           # via requests
web-fragments==0.3.0      # via xblock
webob==1.8.5              # via xblock
xblock==1.2.6             # via edx-when

ua-parser==0.8.0
django-extensions==2.2.9
